Favorites - fixing favorites feature for projects and blinks 09
