+++ app/dashboard/create-listings/projects/page.tsx
@@ -1,7 +1,7 @@
 // /app/dashboard/create-listings/projects/page.tsx
 import { createClient } from "@/utils/supabase/server";
 import { redirect } from "next/navigation";
-import CreateListingsProjects from "@/components/Directory/Dashboard/CreateListings/Projects";
+import CreateListingsProjects from "@/components/Dashboard/CreateListings/Projects";
 import Link from "next/link";
 import Image from "next/image";
 import { Metadata } from "next";
