+++ components/Dashboard/CreateListings/Projects/UploadInfo.tsx
@@ -1,4 +1,4 @@
-// /components/Directory/Dashboard/CreateListings/Blinks/UploadInfo.tsx
+// /components/Dashboard/CreateListings/Projects/UploadInfo.tsx
 import React from 'react';
 
 interface UploadInfoProps {
@@ -39,7 +39,7 @@ const UploadInfo: React.FC<UploadInfoProps> = ({
       <div className="flex flex-col items-center justify-center pt-5 pb-6">
         <p>Drag and drop listing image here, or click to select files</p>
         <em>Only *.jpeg and *.png images will be accepted.</em>
-        <em>MUST be 800x800px (HeightxWidth) Precisely and less than 750KB.</em>
+        <em>MUST be 1200x630px (HeightxWidth) Precisely and less than 750KB.</em>
       </div>
       <input
         ref={fileInputRef}
